+++ src/test/java/com/maxmind/geoip2/model/CountryResponseTest.java
@@ -50,6 +50,7 @@ public class CountryResponseTest {
                     "      \"names\" : {" +
                     "         \"en\" : \"United Kingdom\"" +
                     "      }," +
+                    "      \"is_in_european_union\": true," +
                     "      \"iso_code\" : \"GB\"" +
                     "   }" +
                     "}";
@@ -89,7 +90,9 @@ public class CountryResponseTest {
     @SuppressWarnings("boxing")
     @Test
     public void testCountry() {
-
+        assertEquals(
+                "country.getCountry().isInEuropeanUnion() does not return false",
+                false, this.country.getCountry().isInEuropeanUnion());
         assertEquals("country.getCountry().getCode() does not return US", "US",
                 this.country.getCountry().getIsoCode());
         assertEquals("country.getCountry().getGeoNameId() does not return 1",
@@ -105,6 +108,9 @@ public class CountryResponseTest {
     @Test
     public void testRegisteredCountry() {
         assertEquals(
+                "country.getRegisteredCountry().isInEuropeanUnion() does not return false",
+                false, this.country.getRegisteredCountry().isInEuropeanUnion());
+        assertEquals(
                 "country.getRegisteredCountry().getIsoCode() does not return CA",
                 "CA", this.country.getRegisteredCountry().getIsoCode());
         assertEquals(
@@ -118,6 +124,10 @@ public class CountryResponseTest {
     @SuppressWarnings("boxing")
     @Test
     public void testRepresentedCountry() {
+        assertEquals(
+                "country.getRepresentedCountry().isInEuropeanUnion() does not return true",
+                true,
+                this.country.getRepresentedCountry().isInEuropeanUnion());
         assertEquals(
                 "country.getRepresentedCountry().getCode() does not return GA",
                 "GB", this.country.getRepresentedCountry().getIsoCode());